refactor: 결과 연산 기능 분리
- 각각의 결과를 연산하는 과정을 클래스로 분리함.
