+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,18 +1,15 @@
 package racingcar;
 
-import java.util.List;
-import java.util.stream.IntStream;
 import racingcar.controller.display.DisplayTotalResultController;
 import racingcar.controller.display.DisplayWinnerController;
+import racingcar.controller.process.CalculateTotalResultController;
+import racingcar.controller.process.findWinnersController;
 import racingcar.controller.request.RequestAttemptCountController;
 import racingcar.controller.request.RequestCarsController;
-import racingcar.dto.RoundResult;
 import racingcar.dto.TotalResult;
 import racingcar.dto.Winners;
 import racingcar.model.AttemptCount;
-import racingcar.model.Car;
 import racingcar.model.Cars;
-import racingcar.util.WinnersNameExtractor;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -31,7 +28,7 @@ public class RacingGameApplication {
         AttemptCount attemptCount = setAttemptCount();
 
         TotalResult totalResult = playGame(cars, attemptCount);
-        Winners winners = getWinners(cars);
+        Winners winners = findWinners(cars);
 
         displayTotalResult(totalResult);
         displayWinners(winners);
@@ -46,24 +43,11 @@ public class RacingGameApplication {
     }
 
     private TotalResult playGame(Cars cars, AttemptCount attemptCount) {
-        List<RoundResult> totalResult = IntStream.range(0, attemptCount.count())
-                .mapToObj(attempt -> playRound(cars))
-                .toList();
-
-        return new TotalResult(totalResult);
+        return new CalculateTotalResultController(cars, attemptCount).proceed();
     }
 
-    private RoundResult playRound(Cars cars) {
-        cars.attemptForward();
-        return cars.getRoundResult();
-    }
-
-    private Winners getWinners(Cars cars) {
-        return toWinners(cars.findWinningCars());
-    }
-
-    private Winners toWinners(List<Car> winnersCar) {
-        return WinnersNameExtractor.INSTANCE.extractName(winnersCar);
+    private Winners findWinners(Cars cars) {
+        return new findWinnersController(cars).proceed();
     }
 
     private void displayTotalResult(TotalResult totalResult) {
