+++ src/main/java/racingcar/controller/process/CalculateTotalResultController.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.controller.process;
 
 import java.util.List;
 import java.util.stream.IntStream;
@@ -7,8 +7,16 @@ import racingcar.dto.TotalResult;
 import racingcar.model.AttemptCount;
 import racingcar.model.Cars;
 
-public class RacingGameService {
-    public TotalResult playRounds(Cars cars, AttemptCount attemptCount) {
+public class CalculateTotalResultController {
+    private final Cars cars;
+    private final AttemptCount attemptCount;
+
+    public CalculateTotalResultController(Cars cars, AttemptCount attemptCount) {
+        this.cars = cars;
+        this.attemptCount = attemptCount;
+    }
+
+    public TotalResult proceed() {
         List<RoundResult> totalResult = IntStream.range(0, attemptCount.count())
                 .mapToObj(attempt -> playRound(cars))
                 .toList();
