--- src/main/java/racingcar/dto/AttemptCount.java
@@ -1,4 +1,0 @@
-package racingcar.dto;
-
-public record AttemptCount(int count) {
-}
